+++ tests/StreamTransportTest.php
@@ -28,4 +28,39 @@ class StreamTransportTest extends TransportTestCase
 
         parent::testFollowLocation();
     }
+
+    public function testComposeContextOptions()
+    {
+        $transport = $this->createClient()->getTransport();
+
+        $options = [
+            'protocolVersion' => '1.1',
+            'timeout' => 12,
+            'proxy' => 'tcp://proxy.example.com:5100',
+            'userAgent' => 'Test User Agent',
+            'followLocation' => true,
+            'maxRedirects' => 11,
+            'sslVerifyPeer' => true,
+            'sslCafile' => '/path/to/some/file',
+            'sslCapath' => '/some/path',
+        ];
+        $contextOptions = $this->invoke($transport, 'composeContextOptions', [$options]);
+
+        $expectedContextOptions = [
+            'http' => [
+                'protocol_version' => $options['protocolVersion'],
+                'timeout' => $options['timeout'],
+                'proxy' => $options['proxy'],
+                'user_agent' => $options['userAgent'],
+                'follow_location' => $options['followLocation'],
+                'max_redirects' => $options['maxRedirects'],
+            ],
+            'ssl' => [
+                'verify_peer' => $options['sslVerifyPeer'],
+                'cafile' => $options['sslCafile'],
+                'capath' => $options['sslCapath'],
+            ],
+        ];
+        $this->assertEquals($expectedContextOptions, $contextOptions);
+    }
 }