+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Yii Framework 2 HTTP client extension Change Log
 - Enh #43: Events `EVENT_BEFORE_SEND` and `EVENT_AFTER_SEND` added to `Request` and `Client` (klimov-paul)
 - Enh #46: Added `Request::getFullUrl()` allowing getting the full actual request URL (klimov-paul)
 - Enh #47: Added `Message::addData()` allowing addition of the content data to already existing one (klimov-paul)
+- Enh #50: Option 'protocolVersion' added to `Request::options` allowing specification of the HTTP protocol version (klimov-paul)
 - Enh: Added `XmlFormatter::useTraversableAsArray` allowing processing `\Traversable` as array (klimov-paul)